Organize templates and config code
