+++ python/etl/config/__init__.py
@@ -9,7 +9,6 @@ We use the term "config" files to refer to all files that may reside in the "con
 
 import datetime
 import logging
-import logging.config
 import os
 import os.path
 import re
@@ -25,7 +24,6 @@ import yaml
 from simplejson.errors import JSONDecodeError
 
 import etl.config.dw
-import etl.monitor
 from etl.config.dw import DataWarehouseConfig
 from etl.errors import ETLRuntimeError, InvalidArgumentError, SchemaInvalidError, SchemaValidationError
 
@@ -38,7 +36,7 @@ _dw_config: Optional[DataWarehouseConfig] = None
 _mapped_config: Optional[Dict[str, str]] = None
 
 # Local temp directory used for bootstrap, temp files, etc.
-# TODO(tom): This is a misnomer -- it's also the install dir on EC2 hosts.
+# TODO(tom): This is a misnomer -- it's also the install directory on EC2 hosts.
 ETL_TMP_DIR = "/tmp/redshift_etl"
 
 
@@ -135,30 +133,6 @@ def _build_config_map(settings):
 def etl_tmp_dir(path: str) -> str:
     """Return the absolute path within the ETL runtime directory for the selected path."""
     return os.path.join(ETL_TMP_DIR, path)
-
-
-def configure_logging(full_format: bool = False, log_level: str = None) -> None:
-    """
-    Set up logging to go to console and application log file.
-
-    If full_format is True, then use the terribly verbose format of
-    the application log file also for the console.  And log at the DEBUG level.
-    Otherwise, you can choose the log level by passing one in.
-    """
-    config = load_json("logging.json")
-    if full_format:
-        config["formatters"]["console"] = dict(config["formatters"]["file"])
-        config["handlers"]["console"]["level"] = logging.DEBUG
-    elif log_level:
-        config["handlers"]["console"]["level"] = log_level
-    logging.config.dictConfig(config)
-    # Ignored due to lack of stub in type checking library
-    logging.captureWarnings(True)  # type: ignore
-    logger.info("Starting log for %s with ETL ID %s", package_version(), etl.monitor.Monitor.etl_id)
-    logger.info('Command line: "%s"', " ".join(sys.argv))
-    logger.debug("Current working directory: '%s'", os.getcwd())
-    logger.info(get_release_info())
-    logger.debug(get_python_info())
 
 
 def load_environ_file(filename: str) -> None:
