+++ python/etl/config/default_settings.yaml
@@ -65,7 +65,7 @@
   # Type information from source (PostgreSQL) to target (Redshift)
   "type_maps": {
     # Types that may be used "as-is"
-    # See also # http://docs.aws.amazon.com/redshift/latest/dg/c_Supported_data_types.html
+    # See also http://docs.aws.amazon.com/redshift/latest/dg/c_Supported_data_types.html
     # The keys are regular expression (with escaped backslashes!) and the values are
     # serialization formats (logical types).
     "as_is_att_type": {
