+++ python/etl/templates/__init__.py
@@ -10,12 +10,11 @@ There are two families of templates: text and sql [*].
 
 [*] We play both flavors of music here, Country and Western.
 """
-import fnmatch
 import logging
 import os.path
 import string
 from collections import OrderedDict
-from typing import Dict, List, Optional
+from typing import Dict
 
 import pkg_resources
 import simplejson as json
@@ -117,38 +116,3 @@ def render_sql(template_name: str) -> str:
     """Return SQL query after filling in template."""
     stmt = render_string(template_name, "sql")
     return etl.text.whitespace_cleanup(stmt).rstrip(";")
-
-
-def show_value(name: str, default: Optional[str]) -> None:
-    """
-    Show value of a specific variable.
-
-    This fails if the variable is not set and no default is provided.
-    """
-    value = etl.config.get_config_value(name, default)
-    if value is None:
-        raise InvalidArgumentError("setting '{}' has no value".format(name))
-    print(value)
-
-
-def show_vars(names: List[str]) -> None:
-    """
-    List "variables" with values.
-
-    This shows all known configuration settings as "variables" with their values or just
-    the variables that are selected.
-    """
-    config_mapping = etl.config.get_config_map()
-    all_keys = sorted(config_mapping)
-    if not names:
-        keys = all_keys
-    else:
-        selected_keys = set()
-        for name in names:
-            matching_keys = [key for key in all_keys if fnmatch.fnmatch(key, name)]
-            if not matching_keys:
-                raise InvalidArgumentError("no matching setting for '{}'".format(name))
-            selected_keys.update(matching_keys)
-        keys = sorted(selected_keys)
-    values = [config_mapping[key] for key in keys]
-    print(etl.text.format_lines(zip(keys, values), header_row=["Name", "Value"]))
